Update logic and moved it to the contract name component
